use a hook for navigate instead of window.location w/ burnettk
